+++ mlflow/db.py
@@ -1,7 +1,5 @@
 import click
 
-import mlflow.store.db.utils
-
 
 @click.group("db")
 def commands():
@@ -24,6 +22,8 @@ def upgrade(url):
     large migrations and includes information about how to estimate their performance and
     recover from failures.
     """
+    import mlflow.store.db.utils
+
     engine = mlflow.store.db.utils.create_sqlalchemy_engine_with_retry(url)
     if mlflow.store.db.utils._is_initialized_before_mlflow_1(engine):
         mlflow.store.db.utils._upgrade_db_initialized_before_mlflow_1(engine)
