+++ mlflow/deployments/cli.py
@@ -4,8 +4,7 @@ import json
 import pandas as pd
 from mlflow.utils import cli_args
 from mlflow.deployments import interface
-from mlflow.pyfunc.scoring_server import _get_jsonable_obj
-from mlflow.utils.proto_json_utils import NumpyEncoder
+from mlflow.utils.proto_json_utils import NumpyEncoder, _get_jsonable_obj
 
 
 def _user_args_to_dict(user_list):
@@ -69,7 +68,7 @@ parse_custom_arguments = click.option(
 )
 
 parse_input = click.option(
-    "--input-path", "-I", required=True, help="Path to input json file for prediction",
+    "--input-path", "-I", required=True, help="Path to input json file for prediction"
 )
 
 parse_output = click.option(
